API/requirements.txt
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.0
Flask-Testing==0.6.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyJWT==1.5.3
SQLAlchemy==1.1.14
Werkzeug==0.12.2
